Found and fixed error in Program when viewing the ProfileActivity as a new user
Program would crash if you as a new user without pre-existing
information in the database would try to view the ProfileActivity. in
the addValueEventListener it would try to add values to the editText
objects, values it would "recieve" from an object with no values in it,
thus returning a null.  Fixed the issue by wrapping the block of code in
an if statement.